CLI Parser - because the lua-based config idea wasn't such a good idea (#79)
* CLI Parser - because the lua-based config idea wasn't such a good idea

* isalnum(int) needs unsigned input, otherwise UTF8 and Ansi chars throw an assertion.

Typical problem: (char)255 == (char)-1 but ((char)255) != (int)-1
When isalnum((char)255) is executed, it gets converted into (int)-1, or UINT_MAX (429496...ULL)
This assertion is probably specific to the Microsoft CRT version of isalnum, also it doesn't seem to be a problem in tolower / toupper.

* Removed '--' prefix requirement, strip whitespace when parsing cli.txt files.

+++ CraftEngine/ajek-framework/src/x-stdlib.cpp
@@ -283,7 +283,7 @@ bool xCreateDirectory( const xUniPath& orig_unix_dir )
         tokenSkip = 2;
         pos = 2;
     }
-    if(TARGET_MSW && dir[0] == '/' && isalnum(dir[1])) {
+    if(TARGET_MSW && dir[0] == '/' && isalnum((u8)dir[1])) {
         tokenSkip = 1;
         pos = 2;
     }
